Consolidate the sound normalization into one standalone module
Tighten the three scattered transform modules (rr_phantom, rr_phantom2,
rr_valuesplit) into a single standalone module next to abbreviate:
dyna/analyze/range_restriction_normalize.py.

Drops the superseded flat-phantom transform class (the path-based projection
supersedes it); keeps the flat phantom analysis helpers the value-split needs.
One coherent module with four layers:
  phantom_paths            - sound + invariant projectability analysis
  PhantomProjection        - path-based projection (sees through wrapping)
  ValueSplit               - default+exception for constant overlaps
  RangeRestrictionNormalizer - entry: value-split then project

abbreviate, the refined check (analyze/range_restriction.py), and the spec
transform (transform/range_restriction.py, still the canonical entry point) are
untouched. Behavior preserved: 300-program wrapped/diagonal re-fuzz 0 mismatch;
value-split spot-checks exact. Three test files merged into one
(test_range_restriction_normalize.py). Doc module map (8.2) updated. Suite: 336.
